backend/requirements.txt
# FastAPI and Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.4.0
python-dotenv>=1.0.0
motor>=3.3.0
pymongo>=4.5.0
google-generativeai>=0.3.0
youtube-transcript-api>=0.6.0
requests>=2.31.0
aiofiles>=23.2.0
pydantic-settings==2.1.0

# Database
motor==3.3.2
pymongo==4.6.0
redis==5.0.1

# AI and ML
openai==1.3.0
google-generativeai==0.3.2

# YouTube Data
youtube-comment-downloader==0.1.68
aiohttp==3.9.0
google-api-python-client==2.110.0

# Data Processing
pandas==2.1.4
numpy==1.25.2

# Async and Concurrency
celery==5.3.4
asyncio-throttle==1.0.2

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6
email-validator==2.1.0
validators==0.22.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Logging and Monitoring
structlog==23.2.0
